% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smesir_forecast.R
\name{smesir_forecast}
\alias{smesir_forecast}
\title{smesir_forecast:}
\usage{
smesir_forecast(Jf, model_fit, new_x = NULL)
}
\arguments{
\item{Jf}{Integer length of the forecast time horizon}

\item{model_fit}{A list returned from the \code{smesir} function}

\item{new_x}{An list containing covariate data with which to forecast, as needed by
by the model fit in the antecedent "model_fit" call.}
}
\description{
Forecasts incidence events within the smeSIR model framework. A companion
function to \code{smesir}.
}
